Radial analysis

+++ lib/transformations.py
@@ -49,9 +49,9 @@ def rotate_vol(vol, angles=(0, 0, 0)):
     sx, sy, sz = vol.shape
     sx, sy, sz = (int(sx / 2), int(sy / 2), int(sz / 2))
 
-    offset = np.array([sx, sy, sz], dtype=np.float64)
+    offset = np.array([sx, sy, sz], dtype=np.float32)
     offset = np.dot(mat, offset)
-    tmp = np.array([sx, sy, sz], dtype=np.float64)
+    tmp = np.array([sx, sy, sz], dtype=np.float32)
     offset = tmp - offset
 
     # Apply affine transform
